Restructure dependencies & consolidate compressor
