+++ chia/types/spend_bundle.py
@@ -10,7 +10,7 @@ from clvm.casts import int_from_bytes
 from chia.consensus.default_constants import DEFAULT_CONSTANTS
 from chia.types.blockchain_format.coin import Coin
 from chia.types.blockchain_format.sized_bytes import bytes32
-from chia.util.puzzle_compression import CompressorVersion
+from chia.wallet.util.puzzle_compression import PuzzleCompressor
 from chia.util.streamable import Streamable, dataclass_from_dict, recurse_jsonify, streamable
 from chia.wallet.util.debug_spend_bundle import debug_spend_bundle
 from .blockchain_format.program import Program
@@ -87,16 +87,16 @@ class SpendBundle(Streamable):
     """
 
     @classmethod
-    def compress(cls, bundle: "SpendBundle", version=CompressorVersion()) -> "SpendBundle":
-        compressed_spends: List[CoinSpend] = [CoinSpend.compress(cs, version=version) for cs in bundle.coin_spends]
+    def compress(cls, bundle: "SpendBundle", compressor=PuzzleCompressor()) -> "SpendBundle":
+        compressed_spends: List[CoinSpend] = [CoinSpend.compress(cs, compressor=compressor) for cs in bundle.coin_spends]
         return cls(
             compressed_spends,
             bundle.aggregated_signature,
         )
 
     @classmethod
-    def decompress(cls, bundle: "SpendBundle", version=CompressorVersion()) -> "SpendBundle":
-        decompressed_spends: List[CoinSpend] = [CoinSpend.decompress(cs, version=version) for cs in bundle.coin_spends]
+    def decompress(cls, bundle: "SpendBundle", compressor=PuzzleCompressor()) -> "SpendBundle":
+        decompressed_spends: List[CoinSpend] = [CoinSpend.decompress(cs, compressor=compressor) for cs in bundle.coin_spends]
         return cls(
             decompressed_spends,
             bundle.aggregated_signature,
