+++ chia/wallet/util/puzzle_compression.py
@@ -2,8 +2,8 @@ from typing import Dict, Tuple, List, Union, Any
 
 from chia.types.blockchain_format.program import Program
 from chia.types.blockchain_format.sized_bytes import bytes32
-
 from chia.util.ints import uint16, uint32
+from chia.wallet.util.puzzle_representation import PuzzleRepresentation
 
 # These imports should be temporary
 from blspy import G1Element
@@ -13,73 +13,6 @@ from chia.wallet.puzzles.cc_loader import CC_MOD
 from chia.wallet.puzzles.load_clvm import load_clvm
 
 OFFER_MOD = load_clvm("settlement_payments.clvm")
-
-
-class PuzzleRepresentation:
-    """
-    This class is intended to be vehicle for compression and serialization of an arbitrary configuration of
-    "outer" and "inner" puzzles.  This can act as a generalized "driver" for an arbitrary puzle.
-    The idea is that you make a complex stack of puzzles, represent it with this class, and then you don't have
-    to worry about writing a whole new driver file for your unique stack, you can instead rely on the component
-    drivers to work together as if it were a driver for the whole puzzle.
-    """
-
-    base: bytes32
-    args: List[Union["PuzzleRepresentation", Program]]
-
-    def __init__(self, base, args):
-        self.base = base
-        self.args = args
-
-    def construct(self, driver_dict) -> Program:
-        driver = driver_dict[self.base]
-        return driver.construct(driver_dict, self.args)
-
-    def solve(self, driver_dict, solution_dict: Dict[str, str]) -> Program:
-        driver = driver_dict[self.base]
-        return driver.solve(driver_dict, self.args, solution_dict)
-
-    @classmethod
-    def parse(cls, f) -> "PuzzleRepresentation":
-        blob_len = int.from_bytes(f.read(4), "big")  # It's a uint32 so we read 4 bytes
-        blob = f.read(blob_len)
-        return cls.from_bytes(blob)
-
-    def stream(self, f):
-        blob = bytes(self)
-        prefixed_blob = bytes.fromhex(bytes(uint32(len(blob))).hex() + blob.hex())
-        f.write(prefixed_blob)
-
-    def __bytes__(self) -> bytes:
-        total_bytes = self.base.hex()
-        for arg in self.args:
-            byte_output = bytes(arg)
-            if type(arg) == type(self):
-                total_bytes += "00"
-            else:
-                total_bytes += "01"
-            total_bytes += bytes(uint32(len(byte_output))).hex()
-            total_bytes += byte_output.hex()
-        return bytes.fromhex(total_bytes)
-
-    @classmethod
-    def from_bytes(cls, as_bytes: bytes) -> "PuzzleRepresentation":
-        as_hex = as_bytes.hex()
-        base = bytes32(bytes.fromhex(as_hex[0:64]))
-        as_hex = as_hex[64:]
-        args = []
-        while as_hex != "":
-            num_bytes = int.from_bytes(bytes.fromhex(as_hex[2:10]), "big")
-            end_index = 10 + num_bytes * 2
-            next_bytes = bytes.fromhex(as_hex[10:end_index])
-            if as_hex[0:2] == "00":
-                args.append(cls.from_bytes(next_bytes))
-            else:
-                args.append(Program.from_bytes(next_bytes))
-            as_hex = as_hex[end_index:]
-
-        return cls(base, args)
-
 
 """
 The following are example classes.
@@ -98,7 +31,7 @@ structure below.  They must all have three methods implemented:
 
 class StandardPuzzle:
     @staticmethod
-    def match(puzzle: Program, version: "CompressorVersion") -> Tuple[bool, List[Union[PuzzleRepresentation, Program]]]:
+    def match(puzzle: Program, compressor: "PuzzleCompressor") -> Tuple[bool, List[Union[PuzzleRepresentation, Program]]]:
         uncurried_mod, curried_args = puzzle.uncurry()
         if standard_puzzle.MOD == uncurried_mod:
             synthetic_pubkey = curried_args.first()
@@ -130,12 +63,12 @@ class StandardPuzzle:
 
 class CATPuzzle:
     @staticmethod
-    def match(puzzle: Program, version: "CompressorVersion") -> Tuple[bool, List[Union[PuzzleRepresentation, Program]]]:
+    def match(puzzle: Program, compressor: "PuzzleCompressor") -> Tuple[bool, List[Union[PuzzleRepresentation, Program]]]:
         uncurried_mod, curried_args = puzzle.uncurry()
         if CC_MOD == uncurried_mod:
             tail_hash = curried_args.rest().first()
             innerpuz = curried_args.rest().rest().first()
-            _, matched_inner = PuzzleCompressor.match_puzzle(innerpuz, version=version)
+            _, matched_inner = compressor.match_puzzle(innerpuz)
             return True, [tail_hash, matched_inner]
         else:
             return False, []
@@ -156,7 +89,7 @@ class CATPuzzle:
 
 class OfferPuzzle:
     @staticmethod
-    def match(puzzle: Program, version: "CompressorVersion") -> Tuple[bool, List[Union[PuzzleRepresentation, Program]]]:
+    def match(puzzle: Program, compressor: "PuzzleCompressor") -> Tuple[bool, List[Union[PuzzleRepresentation, Program]]]:
         if OFFER_MOD == puzzle:
             return True, []
         else:
@@ -190,8 +123,10 @@ HASH_TO_DRIVER: Dict[uint16, Dict[bytes32, Any]] = {
 
 LATEST_VERSION: uint16 = uint16(max(HASH_TO_DRIVER.keys()))
 
+class CompressionVersionError(Exception):
+    pass
 
-class CompressorVersion:
+class PuzzleCompressor:
     """
     This class represents an instance of the global HASH_TO_DRIVER map above.
 
@@ -214,8 +149,30 @@ class CompressorVersion:
             self.version_number = uint16(65535)  # Just set it to max so it always passes any version checks
             self.driver_dict = driver_dict
 
+    def match_puzzle(self, puzzle: Program) -> Tuple[bool, Union[PuzzleRepresentation, Program]]:
+        for identifier, driver in self.driver_dict.items():
+            matched, args = driver.match(puzzle, self)
+            if matched:
+                return True, PuzzleRepresentation(identifier, args)
+        return False, puzzle
+
+    def serialize(self, puzzle: Program) -> bytes:
+        _, rep = self.match_puzzle(puzzle)
+        return bytes(self.version_number) + bytes(rep)
+
+    def deserialize(self, object_bytes: bytes) -> Program:
+        if int.from_bytes(object_bytes[0:2], "big") > self.version_number:
+            raise CompressionVersionError()
+        else:
+            deversioned_bytes = object_bytes[2:]
+            try:
+                program = PuzzleRepresentation.from_bytes(deversioned_bytes).construct(self.driver_dict)
+            except Exception:
+                program = Program.from_bytes(deversioned_bytes)
+            return program
+
     @classmethod
-    def get_identifier_version(cls, identifier: bytes32) -> "CompressorVersion":
+    def get_identifier_version(cls, identifier: bytes32) -> "PuzzleCompressor":
         for id, driver_dict in HASH_TO_DRIVER.items():
             if identifier in driver_dict:
                 return cls(id)
@@ -224,7 +181,7 @@ class CompressorVersion:
     @classmethod
     def lowest_compatible_version(
         cls, matched_puzzles: List[Union[PuzzleRepresentation, Program]]
-    ) -> "CompressorVersion":
+    ) -> "PuzzleCompressor":
         highest_version = uint16(0)
         for puz in matched_puzzles:
             if isinstance(puz, PuzzleRepresentation):
@@ -236,36 +193,3 @@ class CompressorVersion:
                     )
                 )
         return cls(highest_version)
-
-
-class CompressionVersionError(Exception):
-    pass
-
-
-class PuzzleCompressor:
-    @classmethod
-    def match_puzzle(
-        cls, puzzle: Program, version=CompressorVersion()
-    ) -> Tuple[bool, Union[PuzzleRepresentation, Program]]:
-        for identifier, driver in version.driver_dict.items():
-            matched, args = driver.match(puzzle, version=version)
-            if matched:
-                return True, PuzzleRepresentation(identifier, args)
-        return False, puzzle
-
-    @classmethod
-    def serialize(cls, puzzle: Program, version=CompressorVersion()) -> bytes:
-        _, rep = cls.match_puzzle(puzzle, version=version)
-        return bytes(version.version_number) + bytes(rep)
-
-    @classmethod
-    def deserialize(cls, object_bytes: bytes, version=CompressorVersion()) -> Program:
-        if int.from_bytes(object_bytes[0:2], "big") > version.version_number:
-            raise CompressionVersionError()
-        else:
-            deversioned_bytes = object_bytes[2:]
-            try:
-                program = PuzzleRepresentation.from_bytes(deversioned_bytes).construct(version.driver_dict)
-            except Exception:
-                program = Program.from_bytes(deversioned_bytes)
-            return program
